requirements.txt
apache-airflow-providers-mongo==4.1.0
apache-airflow-providers-sftp==4.7.0
apache-airflow-providers-jenkins==3.4.0
apache-airflow-providers-apache-spark==4.1.5
apache-airflow-providers-apache-kafka
confluent-kafka==2.1.1
avro-python3=1.10.2
pymemcache=4.0.0
python-dotenv=0.21.1